requirements.txt
Django>=5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-csp>=3.7
whitenoise>=6.6.0
gunicorn>=21.2.0
djangorestframework>=3.16.0
drf-spectacular>=0.28.0
django-permissions-policy>=4.25.0
